+++ swiyu-core-business-service-app/src/main/java/ch/admin/bj/swiyu/core/business/modules/trust/domain/onboarding/DeclarationOfIntent.java
@@ -1,6 +1,6 @@
 package ch.admin.bj.swiyu.core.business.modules.trust.domain.onboarding;
 
-import ch.admin.suis.validator.rest.to.response.FileReport;
+import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
 
-public record DeclarationOfIntent(@NotNull String fullySignedDocumentId, FileReport validationReport) {}
+public record DeclarationOfIntent(@NotNull String fullySignedDocumentId, JsonNode validationReport) {}